Add a basic skeleton of types and traits
Some rudimentary types to represent the domain. Obviously since I'm still
learning Rust all if this is likely to change quite a bit.

First step will be to put together an InMemoryStore for testing out the types
and traits to see what makes sense then more on to more useful implementations.