Add cli frontend skeleton
Begin putting together a CLI tool for the leagus app. My goal will be to have
multiple frontends for the tool; cli, web and desktop. The idea is that the core
will contain all the shared functionality that will allow multiple frontends to
exist with mass duplication.

I'm starting with the CLI as it should be the most straight forward way of
easily getting to a working, but not complete, end to end application. That
should allow me to iterate on the user experience quicker than if I tried to
fully complete each layer in a more horizontal approach.