Add a smidgen more error handling in the mongo store
Since MongoDB is schemaless it is possible to have documents in the collections
which cannot be deserialize into the expected types. I tested that by adding
some bogus docs into the store manually and seeing what happening then trying to
read them.

With this change the program should no longer choke on this documents and
happily ignore them. Later on I might want to log out the docs which fail to
deserialize.