Set the League id as the monogodb id and create a unique index on name
I've renamed the League.id to "_id" using the serde field attribute so mongodb
will use this value as the unique id. On top of that I have added another index
on League.name which is set to unique. Name is not an immutable property but it
must be unique.