Share the monogo client across requests using Axum State extractor
The client itself it shareable across async tasks as it internally tracks refs
via Arc when cloning.

I landed up removing the mutable references in the WriteableStore trait since
they just got in the way and I think I land up removing the trait completely as
it is not providing much value at the moment.