Implement typesafe IDs
The serialised format remains the same as the plain Bson::Uuid but the
deserialized versions are now distinct types. E.g. an ID<Season> is distinct from
an ID<Round> etc...

This should prevent mistakes from assigning the wrong type of ID. This is still
possible when creating the IDs from Uuids but at least that is quite explicit
and should be more obvious when done incorrectly.